refactor search app a bit
