+++ matscholar_web/search/common.py
@@ -1,0 +1,27 @@
+
+import dash_html_components as html
+
+from matscholar_web.common import common_null_warning_html
+
+
+def no_results_html():
+    return common_null_warning_html("No results found!")
+
+
+def results_container_class():
+    return "container has-margin-top-20 has-margin-bottom-20 msweb-fade-in"
+
+
+def get_results_label_html(result_type):
+    if result_type == "entities":
+        label_text = "Statistics (entities)"
+    elif result_type == "materials":
+        label_text = "Summary of Materials"
+    elif result_type == "abstracts":
+        label_text = "Relevant Abstracts"
+    else:
+        raise ValueError(f"Result type {result_type} not valid!")
+
+    label = html.Label(label_text, className="is-size-2 has-margin-10")
+    container = html.Div(label, className="has-margin-top-50")
+    return container
