+++ matscholar_web/search/subviews/abstracts.py
@@ -1,17 +1,12 @@
-from matscholar import Rester
 import dash_html_components as html
-import dash_core_components as dcc
-import json
 import pandas as pd
-import urllib
-from matscholar_web.constants import rester, valid_entity_filters, \
-    entity_shortcode_map, entity_color_map
-from matscholar_web.search.util import parse_search_box, \
-    results_container_class, no_results_html, get_results_label_html
+from matscholar_web.constants import rester, valid_entity_filters, entity_color_map
+from matscholar_web.search.common import results_container_class, no_results_html, get_results_label_html
 
 MAX_N_ABSTRACTS_RETRIEVED = 200  # the number of abstracts retrieved via api
 MAX_N_ABSTRACTS = 20  # the number of abstracts actually shown
 MAX_ENTITIES_PER_ROW = 3
+
 
 def abstracts_results_html(entity_query, raw_text):
     results = rester.abstracts_search(
