+++ matscholar_web/search/subviews/materials.py
@@ -4,7 +4,8 @@ import pandas as pd
 import dash_html_components as html
 
 from matscholar_web.constants import rester
-from matscholar_web.search.util import parse_search_box, no_results_html, \
+from matscholar_web.search.util import parse_search_box
+from matscholar_web.search.common import no_results_html, \
     results_container_class, get_results_label_html
 
 MAX_N_MATERIALS_IN_TABLE = 100
